Fix compile error: textAlign needs explicit uintptr conversion
Extracting the combined DT_* flags into a := variable in
addActionButtons made Go infer type int (the default type for an
untyped constant expression), which then failed to assign into the
drawItem.align field (uintptr) without an explicit conversion — caught
by CI's actual Windows build, since this sandbox has no Go toolchain to
compile against locally.

+++ tray/card.go
@@ -328,7 +328,7 @@ func addActionButtons() {
 	top := cardCursorY
 	cardForceReportRect = winRect{left: s(cardPadX), top: top, right: s(cardPadX) + btnW, bottom: top + h}
 	cardForceEvaluateRect = winRect{left: s(cardPadX) + btnW + gap, top: top, right: s(cardPadX) + btnW + gap + btnW, bottom: top + h}
-	textAlign := dtCenter | dtVCenter | dtSingleLine | dtEndEllipsis
+	textAlign := uintptr(dtCenter | dtVCenter | dtSingleLine | dtEndEllipsis)
 	cardItems = append(cardItems,
 		drawItem{kind: drawKindPill, rect: cardForceReportRect, text: "Force report", font: fontPill, color: cardPrimaryTextColor(cardIsLight), bgColor: cardBorderColor(cardIsLight), align: textAlign, radius: s(8), outline: true},
 		drawItem{kind: drawKindPill, rect: cardForceEvaluateRect, text: "Force evaluate compliance", font: fontPill, color: colWhite, bgColor: colBrand, align: textAlign, radius: s(8)},
